+++ src/Polly.Shared/Utilities/SystemClock.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 namespace Polly.Utilities
 {
     /// <summary>
-    /// Time related delegates used to improve testability of the code
+    /// Time related delegates used to support different compilation targets and to improve testability of the code.
     /// </summary>
     public static class SystemClock
     {